+++ packages/zwave-js/src/Node.ts
@@ -1,3 +1,4 @@
-export type { NODE_ID_BROADCAST, NODE_ID_MAX } from "@zwave-js/core";
+export { NODE_ID_BROADCAST, NODE_ID_MAX } from "@zwave-js/core";
+export { Endpoint } from "./lib/node/Endpoint";
 export { ZWaveNode } from "./lib/node/Node";
-export { NodeStatus } from "./lib/node/Types";
+export { InterviewStage, NodeStatus } from "./lib/node/Types";